3449-maximize-the-minimum-game-score Hard .cpp Array
Solved by: vaibhav_jpd 
LinkedIn: https://www.linkedin.com/in/vaibhav-raj-singh-/

LeetCode: leetcode.com/u/rosvert/

#vaibhav-raj-singh
#jiit
#Jaypee

Student @Jaypee Institute of Information Technology

+++ C++/Hard/3449-maximize-the-minimum-game-score/3449-maximize-the-minimum-game-score.cpp
@@ -25,7 +25,7 @@ private:
         arr[n-1] += nums[n-1];
         m--;
         if(arr[n-1] < mid){
-            int k = (mid - arr[n - 1] + nums[n - 1] - 1) / nums[n - 1];
+            long long int k = (mid - arr[n - 1] + nums[n - 1] - 1) / nums[n - 1];
             return (m >= 2*k);
         }
         return true;
